very rough placement structure
generally using flex

the big margins are present, but additional indentations and
whitespace (e.g. around quote, and signup) are not accounted for.

and spacing is extremely rough and unfinished.

and not all styling is complete.

consider this a first draft, pointing it in the rough direction of the
page look and feel, but many more details are to come.