Fix inclusion of Google Roboto fonts
I think I was getting the regular, but not the others.

I got the updated HTML from going to here:
* https://fonts.google.com/specimen/Roboto#styles
And selecting these styles:
* Light 300 Italic
* Regular 400
* Bold 700

Then copying from "Use on the web" by checking "link"

Note that there are other styles available

To check if they are being properly included, you can change the
fallback in "font-family" in "body" in the CSS from sans-serif to
serif, which makes it obvious to see when it's wrong.